requirements.txt
Flask
pandas
openpyxl
python-docx
PyPDF2
Pillow
matplotlib
wordcloud
requests
